fix: export color picker
