+++ src/components/ColorPicker/index.ts
@@ -1,0 +1,2 @@
+export { VColorPicker } from './ColorPicker';
+export { TypePickerColor,VColorResult} from'./types'
