+++ src/components/index.ts
@@ -37,3 +37,4 @@ export {
 export { IVBadgeProps, IVBadgeStyle, VBadge } from './Badge';
 export { VLoadProgressTask } from './LoadProgressTask';
 export { showToastNotification } from './ToastNotification';
+export {VColorPicker,VColorResult,TypePickerColor } from './ColorPicker'
